refactor(client): differentiate player reference setting from ready state
